build: Improve FreeType dep search
FreeType supported building with Meson only since 2.10.3, and had different DLL
naming from the Visual Studio projects and CMake build files (and the
pkg-config files generated with the CMake build system may not be usable
out-of-the-box), so try finding FreeType using pkg-config first, and then try
using CMake.  Look for FT_Get_Color_Glyph_Paint() according as a result.

Sadly, we can't use the multiple dependency naming mechanism here because the
pkg-config files use the libtool versioning scheme for FreeType and CMake looks
for the actual version of FreeType (which does the locating of headers and
libraries under the hood via paths in %INLUDE% and %LIB% for Visual Studio-like
builds).